refactor Bind structure & build config for v1.3.2
Removed Win32 platform support and solution/filter files from Bind. Updated Bind.vcxproj to use explicit x64 Qt/Vcpkg paths, set new output directories, and simplified build settings. Changed GitHub Actions workflow to match new build output locations. Migrated resource and artifact paths, and added missing headers for improved modularity.

+++ Bind/src/Core/Utils/Utils.cpp
@@ -1,20 +1,6 @@
-#include "include/Core/Utils/Utils.h"
-#include "include/Core/Utils/PathUtils.h"
-#include "include/Core/Obfuscation/Obfuscation.h"
-#include "include/Core/Obfuscation/Direct/Encryption/DirectEncryptor.h"
-#include <QDebug>
-#include <QByteArray>
-#include <cstring>
-#include <QFile>
-#include <QCryptographicHash>
-#include <QJsonDocument>
-#include <QJsonObject>
-#include <QJsonArray>
-#include <QRegularExpression>
-#include <QRegularExpressionMatch>
-#include <QDateTime>
-#include <QDir>
-#include <QSettings>
+#include <Core/Obfuscation/Obfuscation.h>
+#include <Core/Obfuscation/Direct/Direct.h>
+#include <Core/Utils/Common.h>
 
 QMap<QString, int> SyscallExtractor::getSyscallsFromDll(const QString& dllPath)
 {
@@ -62,7 +48,7 @@ QMap<QString, int> SyscallExtractor::getSyscallsFromDll(const QString& dllPath)
 
         try
         {
-            funcName = QString::fromUtf8(fn.c_str(), fn.length());
+            funcName = QString::fromUtf8(fn.c_str(), static_cast<int>(fn.length()));
         }
         catch (...)
         {
